refactor: colocate route helpers
