+++ app/routes/-helpers/api/hono-client.ts
@@ -1,5 +1,5 @@
 import { hc } from "hono/client"
-import type { AppType } from "../../routes/api"
+import type { AppType } from "../../api"
 
 export const createHonoClient = () => {
   return hc<AppType>("/api")
