+++ app/routes/_renderer.tsx
@@ -1,9 +1,9 @@
 import { Style } from "hono/css"
 import { jsxRenderer, useRequestContext } from "hono/jsx-renderer"
 import { Link, Script } from "honox/server"
-import colorSchemeInitScript from "../helpers/ui/color-scheme-entry?js"
-import { deleteToastCookie, getToastCookie } from "../helpers/ui/toast"
 import Toast from "./-components/ui/toast"
+import colorSchemeInitScript from "./-helpers/ui/color-scheme-entry?js"
+import { deleteToastCookie, getToastCookie } from "./-helpers/ui/toast"
 import staffSidebarInitScript from "./staff/-helpers/sidebar-entry?js"
 
 export default jsxRenderer(({ children }) => {
