+++ app/routes/staff/products/[id]/delete/index.tsx
@@ -1,5 +1,4 @@
 import { createRoute } from "honox/factory"
-import { setToastCookie } from "../../../../../helpers/ui/toast"
 import { removeProduct } from "../../../../../usecases/commands/removeProduct"
 import { getProductDeletePageData } from "../../../../../usecases/queries/getProductDeletePageData"
 import { formatCurrencyJPY } from "../../../../../utils/money"
@@ -8,6 +7,7 @@ import Trash2Icon from "../../../../-components/icons/lucide/trash2Icon"
 import Button from "../../../../-components/ui/button"
 import Chip from "../../../../-components/ui/chip"
 import LinkButton from "../../../../-components/ui/linkButton"
+import { setToastCookie } from "../../../../-helpers/ui/toast"
 import Layout from "../../../-components/layout"
 
 export const POST = createRoute(async (c) => {
