+++ app/routes/staff/products/[id]/edit/index.tsx
@@ -1,8 +1,8 @@
 import { validator } from "hono/validator"
 import { createRoute } from "honox/factory"
-import { setToastCookie } from "../../../../../helpers/ui/toast"
 import { registerProduct } from "../../../../../usecases/commands/registerProduct"
 import { getProductEditPageData } from "../../../../../usecases/queries/getProductEditPageData"
+import { setToastCookie } from "../../../../-helpers/ui/toast"
 import Layout from "../../../-components/layout"
 import ProductRegistrationForm from "../../-components/$productRegistrationForm"
 import { parseProductRegistrationFormData } from "../../-helpers/parseProductRegistrationFormData"
