+++ app/routes/staff/products/index.tsx
@@ -2,7 +2,6 @@ import type { FC } from "hono/jsx"
 import { validator } from "hono/validator"
 import { createRoute } from "honox/factory"
 import { tv } from "tailwind-variants"
-import { setToastCookie } from "../../../helpers/ui/toast"
 import { registerProduct } from "../../../usecases/commands/registerProduct"
 import {
   getProductsManagementPageData,
@@ -16,6 +15,7 @@ import TriangleAlertIcon from "../../-components/icons/lucide/triangleAlertIcon"
 import Badge from "../../-components/ui/badge"
 import Chip from "../../-components/ui/chip"
 import ItemCollectionViewer from "../../-components/ui/itemCollectionViewer"
+import { setToastCookie } from "../../-helpers/ui/toast"
 import Layout from "../-components/layout"
 import ProductRegistrationForm from "./-components/$productRegistrationForm"
 import { parseProductRegistrationFormData } from "./-helpers/parseProductRegistrationFormData"
